requirements.txt
# Python dependencies for the Reddit Problem Validation System

# Core Framework (choose one)
# flask
django

# Reddit API Wrapper
praw

# Web Scraping
beautifulsoup4
scrapy
requests

# AI/ML (choose based on preference/need)
tensorflow
torch>=2.0.0
transformers
tqdm>=4.65.0
sentence-transformers>=2.2.2
scikit-learn
nltk
spacy
tf-keras>=2.15.0

# Database (choose one or both)
pymongo
psycopg2-binary

# API Development
# flask-restful  # If using Flask
djangorestframework # If using Django

# Utilities
python-dotenv # For managing environment variables
pandas # For data manipulation and CSV output
